Move support functions to env support file
