+++ gym-drill/gym_drill/envs/drill_env.py
@@ -10,6 +10,7 @@ from gym_drill.envs.Coordinate import Coordinate
 from gym_drill.envs.ObservationSpace import ObservationSpace
 from gym_drill.envs.Target import TargetBall
 from gym_drill.envs.Hazard import Hazard
+from gym_drill.envs import environment_support as es
 
 # Max values for angular velocity and acceleration
 MAX_ANGVEL = 0.1
@@ -79,11 +80,11 @@ class DrillEnv(gym.Env):
         self.initialAngAcc = bitInitialization[2]
 
         # Init targets. See _init_targets function
-        self.targets = _init_targets(NUM_TARGETS,TARGET_BOUND_X,TARGET_BOUND_Y,TARGET_RADII_BOUND,startLocation)
+        self.targets = es._init_targets(NUM_TARGETS,TARGET_BOUND_X,TARGET_BOUND_Y,TARGET_RADII_BOUND,startLocation)
         self.activate_hazards = activate_hazards
         if self.activate_hazards:
             #print("Initiating environment with hazards")
-            self.hazards = _init_hazards(NUM_HAZARDS,HAZARD_BOUND_X,HAZARD_BOUND_Y,HAZARD_RADII_BOUND,startLocation,self.targets)
+            self.hazards = es._init_hazards(NUM_HAZARDS,HAZARD_BOUND_X,HAZARD_BOUND_Y,HAZARD_RADII_BOUND,startLocation,self.targets)
         else:
             #print("Initiating environment without hazards")
             self.hazards = []
@@ -101,7 +102,7 @@ class DrillEnv(gym.Env):
         # Log related
         self.episode_counter = 0 # Used to write to log
         self.total_reward = 0      
-        _init_log()
+        es._init_log()
 
     def seed(self, seed=None):
         self.np_random, seed = seeding.np_random(seed)
@@ -128,14 +129,15 @@ class DrillEnv(gym.Env):
         # If drill is no longer on screen, game over.
         if not (0 < self.bitLocation.x < SCREEN_X and 0 < self.bitLocation.y < SCREEN_Y):
             reward  -=30
-            done = True     
+            done = True   
         
-        """
+        
         # Check if we hit a hazard
         for h in self.hazards:
-            if _is_within(self.bitLocation,h.center,h.radius):
-                reward -= 1000.0
-        """
+            if es._is_within(self.bitLocation,h.center,h.radius):
+                reward -= 100.0
+                done = True
+                print("Hazard hit, I will stop")        
 
         if len(self.step_history)>NUM_MAX_STEPS:
             done= True                        
@@ -191,8 +193,6 @@ class DrillEnv(gym.Env):
         angle_between_vectors = np.math.atan2(np.linalg.det([appr_vec, head_vec]), np.dot(appr_vec, head_vec))
 
         return angle_between_vectors
-
-
     
     # For encapsulation. Updates the bit according to the action
     def update_bit(self,action):
@@ -253,12 +253,12 @@ class DrillEnv(gym.Env):
         self.step_history = [[self.start_x,self.start_y]]       
 
         # Need to init new targets
-        self.targets = _init_targets(NUM_TARGETS,TARGET_BOUND_X,TARGET_BOUND_Y,TARGET_RADII_BOUND,self.bitLocation)             
+        self.targets = es._init_targets(NUM_TARGETS,TARGET_BOUND_X,TARGET_BOUND_Y,TARGET_RADII_BOUND,self.bitLocation)             
         
         # Init new hazards
         if self.activate_hazards:
             #print("Initiating environment with hazards")
-            self.hazards = _init_hazards(NUM_HAZARDS,HAZARD_BOUND_X,HAZARD_BOUND_Y,HAZARD_RADII_BOUND,self.bitLocation,self.targets)
+            self.hazards = es._init_hazards(NUM_HAZARDS,HAZARD_BOUND_X,HAZARD_BOUND_Y,HAZARD_RADII_BOUND,self.bitLocation,self.targets)
         else:
             #print("Initiating environment without hazards")
             self.hazards = []
@@ -388,127 +388,22 @@ class DrillEnv(gym.Env):
         return self.viewer.render(return_rgb_array = mode=='rgb_array')
     """
 
-def _init_log(*,filename="drill_log.txt"):
-    f = open(filename,"w")
-    init_msg = "Log for training session started at " + str(datetime.now()) +"\n \n"
-    f.write(init_msg)
-    f.close()
-    #print("Log created!")
-
-
-# Returns an ordered list of randomly generated targets within the bounds given. 
-def _init_targets(num_targets,x_bound,y_bound,r_bound,start_location):
-    all_targets = []
-
-    for t in range(num_targets):
-        target = _create_unique_random_target(start_location,x_bound,y_bound,r_bound,all_targets)
-        all_targets.append(target)        
-    
-    all_targets = _orderTargets(start_location,all_targets)
-
-    return all_targets
-
-def _init_hazards(num_hazards,x_bound,y_bound,r_bound,start_pos,existing_targets):
-    all_hazards = []
-    for h in range(num_hazards):
-        avoid = existing_targets + all_hazards
-        hazard = _create_unique_random_hazard(start_pos,x_bound,y_bound,r_bound,avoid)
-        all_hazards.append(hazard)
-
-    return all_hazards
-
-# Finds nearest between 1 point and a list of candidate points
-# startlocation is type Coordinate, and candidates is list of types Targets
-def _findNearest(start_location,candidates):
-    current_shortest_distance = -1 # Init with an impossible distance
-    current_closest_target_index = 0
-    for candidate_index in range(len(candidates)):        
-        candidate = candidates[candidate_index]     
-        distance = Coordinate.getEuclideanDistance(candidate.center,start_location)
-        
-        if distance < current_shortest_distance or current_shortest_distance == -1:           
-            current_shortest_distance = distance
-            current_closest_target_index = candidate_index
-        
-    return current_closest_target_index
-
-# Orders the target based upon a given start location
-# start_location is type Coordinate, all_targets is list of type targets
-def _orderTargets(start_location,all_targets):
-    #target_order = [None] * len(all_targets) # Maybe better with = [] and use append()
-    target_order = [] 
-    loop_counter = 0    
-
-    while len(all_targets) != 0:
-        if loop_counter == 0:
-            next_in_line_index = _findNearest(start_location,all_targets)
-            next_in_line_target = all_targets[next_in_line_index]
-            target_order.append(next_in_line_target)
-            all_targets.pop(next_in_line_index)
-        else:
-            next_in_line_index = _findNearest(target_order[loop_counter-1].center,all_targets)
-            next_in_line_target = all_targets[next_in_line_index]
-            target_order.append(next_in_line_target)
-            all_targets.pop(next_in_line_index)  
-
-        loop_counter += 1  
-
-    return target_order
-
-def _is_within(bitPosition,targetPosition,targetRadius):
-    return (bitPosition.x - targetPosition.x)**2 + (bitPosition.y - targetPosition.y)**2 < targetRadius
-
-
-# Returns True if t1 or t2 overlap. Works for both Hazards and Targets
-def _is_overlapping(t1,t2):
-    total_radii = t1.radius + t2.radius
-    distance = Coordinate.getEuclideanDistance(t1.center,t2.center)
-    return  distance < total_radii
-
-# Creates a uniqe target that does not overlap with any targets in existing_targets
-def _create_unique_random_target(start_pos,x_bound,y_bound,r_bound,existing_targets):
-    target_center = Coordinate(np.random.uniform(x_bound[0],x_bound[1]),(np.random.uniform(y_bound[0],y_bound[1] )))
-    target_radius = np.random.uniform(r_bound[0],r_bound[1])
-    target_candidate = TargetBall(target_center.x,target_center.y,target_radius)
-
-    for target in existing_targets:
-        if _is_overlapping(target,target_candidate) or _isWithin(start_pos,target_center,target_radius):
-            target_candidate =_create_unique_random_target(start_pos,x_bound,y_bound,r_bound,existing_targets)
-            break
-
-    return target_candidate
-
-# Creates a uniqe hazard that does not overlad with any obstacles in existing_obstacles
-def _create_unique_random_hazard(start_pos,x_bound,y_bound,r_bound,existing_obstacles):
-    hazard_center = Coordinate(np.random.uniform(x_bound[0],x_bound[1]),(np.random.uniform(y_bound[0],y_bound[1] )))
-    hazard_radius = np.random.uniform(r_bound[0],r_bound[1])
-    hazard_candidate = Hazard(hazard_center.x,hazard_center.y,hazard_radius)  
-    
-    for obstacle in existing_obstacles:
-        if _is_overlapping(obstacle,hazard_candidate) or _isWithin(start_pos,hazard_center,hazard_radius):
-            hazard_candidate = _create_unique_random_hazard(start_pos,x_bound,y_bound,r_bound,existing_obstacles)
-            break
-    
-    return hazard_candidate
-
-def _isWithin(bitPosition,targetPosition,targetRadius):
-    return (bitPosition.x - targetPosition.x)**2 + (bitPosition.y - targetPosition.y)**2 < targetRadius
 
 if __name__ == '__main__':
     print("Testing init of targets and hazards")    
     startpos = Coordinate(100,400)
 
     print("Creating targets")
-    t = _init_targets(NUM_TARGETS,TARGET_BOUND_X,TARGET_BOUND_Y,TARGET_RADII_BOUND,startpos)
+    t = es._init_targets(NUM_TARGETS,TARGET_BOUND_X,TARGET_BOUND_Y,TARGET_RADII_BOUND,startpos)
     for _ in t:
         print(_)
     
     print("Creating Hazards")    
-    h = _init_hazards(NUM_HAZARDS,HAZARD_BOUND_X,HAZARD_BOUND_Y,HAZARD_RADII_BOUND,startpos,t)
+    h = es._init_hazards(NUM_HAZARDS,HAZARD_BOUND_X,HAZARD_BOUND_Y,HAZARD_RADII_BOUND,startpos,t)
     for eden_hazard in h:
         print(eden_hazard)
     
-    # Plot circles from targetballs, colors just to verify the order of the balls
+    # plot circles from targetballs, colors just to verify the order of the balls
     theta = np.linspace(0, 2*np.pi, 100)
     colors_order = {
         1:"b",
