+++ gym-drill/main_no_training.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from gym_drill.envs.Coordinate import Coordinate
 
-STARTLOCATION = Coordinate(100,1000)
+STARTLOCATION = Coordinate(150,850)
 BIT_INITIALIZATION = [3.8*np.pi/4,0.0,0.0]
 
 env_name = 'drill-v0'
